refactor: convert service to object
