+++ test/npmrc-auth.ts
@@ -5,17 +5,17 @@ import {
 } from "../src/io/npmrc-io";
 import { AsyncResult, Err, Ok } from "ts-results-es";
 import { RequiredEnvMissingError } from "../src/io/upm-config-io";
-import { tryUpdateUserNpmrcToken } from "../src/services/npmrc-token-update";
 import { emptyNpmrc, setToken } from "../src/domain/npmrc";
 import { exampleRegistryUrl } from "./data-registry";
 import { IOError } from "../src/io/file-io";
+import { makeNpmrcAuthService } from "../src/services/npmrc-auth";
 
 const exampleNpmrcPath = "/users/someuser/.npmrc";
 
 jest.mock("../src/io/npmrc-io");
 
-describe("npmrc-token-update service", () => {
-  describe("user-wide", () => {
+describe("npmrc-auth", () => {
+  describe("set token", () => {
     beforeEach(() => {
       // Mock defaults
       jest.mocked(tryGetNpmrcPath).mockReturnValue(Ok(exampleNpmrcPath));
@@ -25,9 +25,10 @@ describe("npmrc-token-update service", () => {
 
     it("should fail if path could not be determined", async () => {
       const expected = new RequiredEnvMissingError();
+      const service = makeNpmrcAuthService();
       jest.mocked(tryGetNpmrcPath).mockReturnValue(Err(expected));
 
-      const result = await tryUpdateUserNpmrcToken(
+      const result = await service.trySetAuthToken(
         exampleRegistryUrl,
         "some token"
       ).promise;
@@ -37,9 +38,10 @@ describe("npmrc-token-update service", () => {
 
     it("should fail if npmrc load failed", async () => {
       const expected = new IOError();
+      const service = makeNpmrcAuthService();
       jest.mocked(tryLoadNpmrc).mockReturnValue(new AsyncResult(Err(expected)));
 
-      const result = await tryUpdateUserNpmrcToken(
+      const result = await service.trySetAuthToken(
         exampleRegistryUrl,
         "some token"
       ).promise;
@@ -49,9 +51,10 @@ describe("npmrc-token-update service", () => {
 
     it("should fail if npmrc save failed", async () => {
       const expected = new IOError();
+      const service = makeNpmrcAuthService();
       jest.mocked(trySaveNpmrc).mockReturnValue(new AsyncResult(Err(expected)));
 
-      const result = await tryUpdateUserNpmrcToken(
+      const result = await service.trySetAuthToken(
         exampleRegistryUrl,
         "some token"
       ).promise;
@@ -60,7 +63,9 @@ describe("npmrc-token-update service", () => {
     });
 
     it("should return npmrc path", async () => {
-      const result = await tryUpdateUserNpmrcToken(
+      const service = makeNpmrcAuthService();
+
+      const result = await service.trySetAuthToken(
         exampleRegistryUrl,
         "some token"
       ).promise;
@@ -73,8 +78,9 @@ describe("npmrc-token-update service", () => {
     it("should create npmrc if missing", async () => {
       const expected = setToken(emptyNpmrc, exampleRegistryUrl, "some token");
       const saveSpy = jest.mocked(trySaveNpmrc);
+      const service = makeNpmrcAuthService();
 
-      await tryUpdateUserNpmrcToken(exampleRegistryUrl, "some token").promise;
+      await service.trySetAuthToken(exampleRegistryUrl, "some token").promise;
 
       expect(saveSpy).toHaveBeenCalledWith(exampleNpmrcPath, expected);
     });
@@ -88,8 +94,9 @@ describe("npmrc-token-update service", () => {
       const expected = setToken(initial, exampleRegistryUrl, "some token");
       jest.mocked(tryLoadNpmrc).mockReturnValue(new AsyncResult(Ok(initial)));
       const saveSpy = jest.mocked(trySaveNpmrc);
+      const service = makeNpmrcAuthService();
 
-      await tryUpdateUserNpmrcToken(exampleRegistryUrl, "some token").promise;
+      await service.trySetAuthToken(exampleRegistryUrl, "some token").promise;
 
       expect(saveSpy).toHaveBeenCalledWith(exampleNpmrcPath, expected);
     });
